[B] Remove potential production cyclical dependencies
In production, loading behaves slightly differently and there are
interdependencies with `AppTypes` that are causing the application to
fail to load. In development and test, even with eager loading, it
would pass. This moves most all references to AppTypes to the support
GlobalTypes, which contains no application code, or specific type
namespaces in order to avoid cyclical dependencies.

In addition, we remove some old/duplicate operations that have been
supplanted by VOG code in lib/support.