[E] Improve performance of dynamic orderings
Add an EntityCachedAncestor materialized view that is equivalent to
the EntityAncestor view, but cached so as to expedite dynamic query
lookups in the template system. This view is refreshed periodically
in a background job, but may occasionally be out of date. Since
dynamic orderings are evaluated at request time, this should cause
very little disruption, if any.